food module api integration also modularised the file into different sub parts
