+++ client/lib/model/sugar_data_point.dart
@@ -1,0 +1,7 @@
+// Sugar data point class for Syncfusion charts
+class SugarDataPoint {
+  final double time;
+  final double value;
+
+  SugarDataPoint(this.time, this.value);
+}
